fix: macOS/Windows CI + revamp artifact packaging
- Fix narrowing conversion in geometry_test.cpp (int -> uint64_t for tri_t)
- Fix lcov --ignore-errors unused for Ubuntu Debug coverage
- Add CMAKE_CXX_STANDARD=17 for Windows to fix Boost _snprintf
- Remove smoothing.cpp UB (uninitialized vars, leftover from prior fix)
- Replace CPack (DEB/DMG/NSIS) with archive-based packaging:
  - zip for Windows and macOS, tar.gz for Linux
  - Filenames: libcvc-{version}-{commit}-{os}-{arch}-{build_type}
  - PDB files included for Windows Debug builds

+++ src/cvc/tests/geometry_test.cpp
@@ -1756,7 +1756,7 @@ TEST(AlgorithmTest, SDFV2ParallelExecution) {
   geometries[0].points().push_back({{0.5, 0.5, 0.5}});
   geometries[0].points().push_back({{-0.5, 0.5, 0.5}});
   for (int i = 0; i < 6; i++) {
-    int base = i * 2;
+    uint64_t base = static_cast<uint64_t>(i * 2);
     geometries[0].tris().push_back({{base % 8, (base + 1) % 8, (base + 2) % 8}});
     geometries[0].tris().push_back({{base % 8, (base + 2) % 8, (base + 3) % 8}});
   }
